requirements.txt
opencv-python==4.7.0.72
torch==2.0.0
gradio==3.26.0
transformers==4.26.0
Pillow==9.5.0
numpy==1.24.0

